+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+
+* Improve tests for the cli module that interacts with mitmproxy.
+* Guard against non-sudo when proxy-settings is enabled.
+
 ## 1.0.0-beta3
 
 * Fix the build.